1215-1 step4

+++ src/main/java/hexlet/code/service/userService/UserServiceImpl.java
@@ -1,25 +1,26 @@
 package hexlet.code.service.userService;
 
 import hexlet.code.dto.UserDto;
-import hexlet.code.model.QUser;
+import hexlet.code.exception.DefaultException;
+import hexlet.code.exception.ExistsException;
+import hexlet.code.exception.ForbiddenException;
+import hexlet.code.exception.UnprocessibleException;
 import hexlet.code.model.User;
 import hexlet.code.repository.UserRepository;
+import hexlet.code.utils.UserUtils;
+import jakarta.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.server.ResponseStatusException;
 
-import javax.crypto.SecretKeyFactory;
-import javax.crypto.spec.PBEKeySpec;
 import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
 import java.security.spec.InvalidKeySpecException;
-import java.security.spec.KeySpec;
-import java.util.Arrays;
-import java.util.List;
+import java.util.concurrent.ExecutionException;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -30,11 +31,19 @@ public class UserServiceImpl implements UserService {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
+    @Autowired
+    private UserUtils userUtils;
+
     @Override
     public User findById(long id) {
-        return userRepository.findById(id)
+        User user = userRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY,
                         "User with id " + id + " not found")); //после сообщения еще есть экспшен
+        if (!userUtils.currentHasAccessTo(user)) {
+            throw new ForbiddenException("User with id " + id + " is not accessible with current login: "
+                    + userUtils.getCurrentLogin());
+        }
+        return user;
     }
 
     @Override
@@ -43,58 +52,61 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User createUser(UserDto dto) throws NoSuchAlgorithmException, InvalidKeySpecException {
-        User user = new User(dto);
-        user.setPasswordDigest(passwordEncoder.encode(dto.getPassword()));
-        userRepository.save(user);
-        return user;
+    public User createUser(@Valid UserDto dto) {
+        try {
+            User newUser = new User(dto);
+            if (userRepository.findByEmail(newUser.getEmail()).isPresent()) {
+                throw new ExistsException("User with email " + newUser.getEmail() + " already exists");
+            }
+            newUser.setPasswordDigest(passwordEncoder.encode(dto.getPassword()));
+            userRepository.save(newUser);
+            return newUser;
+        } catch (RuntimeException exception) {
+            throw new DefaultException("default msg 418 plus " + exception.getMessage());
+        }
     }
 
     @Override
-    public User updateUser(long id, UserDto dto) throws NoSuchAlgorithmException, InvalidKeySpecException {
-        User toUpdate = userRepository.findById(id)
+    public User updateUser(long id, @Valid UserDto dto) throws NoSuchAlgorithmException, InvalidKeySpecException {
+        User userToUpdate = userRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY,
                         "User with id " + id + " not found")); //после сообщения еще есть экспшен
-        toUpdate.setFirstName(dto.getFirstName());
-        toUpdate.setLastName(dto.getLastName());
-        toUpdate.setEmail(dto.getEmail());
-        toUpdate.setPasswordDigest(passwordEncoder.encode(dto.getPassword()));
-        userRepository.save(toUpdate);
+
+        if (!userUtils.currentHasAccessTo(userToUpdate)) {
+            throw new ForbiddenException("User with id " + id + " is not accessible with current login: "
+                    + userUtils.getCurrentLogin());
+        }
+        userToUpdate.setFirstName(dto.getFirstName());
+        userToUpdate.setLastName(dto.getLastName());
+        userToUpdate.setEmail(dto.getEmail());
+        userToUpdate.setPasswordDigest(passwordEncoder.encode(dto.getPassword()));
+        userRepository.save(userToUpdate);
         return findById(id);
     }
 
     @Override
     public void deleteUser(long id) {
-        User toDelete = userRepository.findById(id)
+        User userToDelete = userRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY,
                         "User with id " + id + " not found")); //после сообщения еще есть экспшен
+        if (!userUtils.currentHasAccessTo(userToDelete)) {
+            throw new ForbiddenException("User with id " + id + " is not accessible with current login: "
+                    + userUtils.getCurrentLogin());
+        }
         userRepository.deleteById(id);
     }
 
-    //no usages:
-    private User findByDto(UserDto dto) {
-        List<User> usersToReturn = (List) userRepository.findAll(
-                QUser.user.firstName.eq(dto.getFirstName())
-                        .and(QUser.user.lastName.eq(dto.getLastName()))
-                        .and(QUser.user.email.eq(dto.getEmail()))
-        );
-        if (usersToReturn.size() != 1) {
-            throw new ResponseStatusException(HttpStatus.MULTIPLE_CHOICES,
-                    "Can't specify User, there are few, with such combination"); //после сообщения еще есть экспшен
-        }
-        return usersToReturn.get(0);
-    }
-    //no usages:
-    private String getHashedPwd(String stringPwd) throws NoSuchAlgorithmException, InvalidKeySpecException {
-        SecureRandom random = new SecureRandom();
-        byte[] salt = new byte[16];
-        random.nextBytes(salt);
-
-        KeySpec spec = new PBEKeySpec(stringPwd.toCharArray(), salt, 2, 32);
-        SecretKeyFactory factory = SecretKeyFactory.getInstance("PBKDF2WithHmacSHA1");
-
-        String hashedPwd = Arrays.toString(factory.generateSecret(spec).getEncoded());
-        log.info("§§§§§§ " + stringPwd + " -> " + hashedPwd);
-        return hashedPwd;
-    }
+//    no usages:
+//    private String getHashedPwd(String stringPwd) throws NoSuchAlgorithmException, InvalidKeySpecException {
+//        SecureRandom random = new SecureRandom();
+//        byte[] salt = new byte[16];
+//        random.nextBytes(salt);
+//
+//        KeySpec spec = new PBEKeySpec(stringPwd.toCharArray(), salt, 2, 32);
+//        SecretKeyFactory factory = SecretKeyFactory.getInstance("PBKDF2WithHmacSHA1");
+//
+//        String hashedPwd = Arrays.toString(factory.generateSecret(spec).getEncoded());
+//        log.info("§§§§§§ " + stringPwd + " -> " + hashedPwd);
+//        return hashedPwd;
+//    }
 }
